Node::lazyAttach should not synchronously attach inside recalcStyle
There was a hack in Node::lazyAttach for <object> since the loader
callback can happen inside ::attach which triggers a reattach to show
the fallback content, but this breaks InsertionPoint::detach calling
lazyReattach since now we'll call attach() for each distributed node
inside the detach of the insertion point (and its ancestor) which means
the nodes are reattached as if their parent was display: none.

To fix this we move the hack for <object> to HTMLObjectElement and make
lazyAttach never synchronously attach.

BUG=277160

Review URL: https://chromiumcodereview.appspot.com/22790006

git-svn-id: svn://svn.chromium.org/blink/trunk@156534 bbb929c8-8fbe-4397-9dbb-9b2b20218538