Remove code related to title directionality, we never used it
The project to support text direction in the page title (and history items) was never completed
and doesn't appear to work once we lazy attach in the parser. It's not clear why the style doesn't
inherit properly down through the <head>, but once we lazy attach the style never gets there.

Some day if Chromium decides to support text directions in titles we can resurrect this code
and figure out how to make it work in a lazy attach world.

BUG=27094

Review URL: https://chromiumcodereview.appspot.com/23190034

git-svn-id: svn://svn.chromium.org/blink/trunk@156607 bbb929c8-8fbe-4397-9dbb-9b2b20218538