Replace HTMLTextFormControlElement::fixPlaceholderRenderer with an override of RenderTextControl::addChild
fixPlaceholderRenderer was used to rearrange the render tree to make sure that
the placeholder text was always first. It required a synchronous attach hook though
so isntead just teach RenderTextControl::addChild how to do this reordering. The hack
is still really ugly, but at least it's contained in one place.

Review URL: https://chromiumcodereview.appspot.com/23035007

git-svn-id: svn://svn.chromium.org/blink/trunk@156619 bbb929c8-8fbe-4397-9dbb-9b2b20218538