Make possible to construct an AnimationEvent.
This implements the possibility to construct 
an AnimationEvent yourself.

This patch doesn't implement yet the dispatching of it.

Note that I had to rename the AnimationEvent class to WebKitAnimationEvent.
This is really unfortunate but I couldn't find a better way. 
EventAliases could potentially support my use case (aliasing the unprefixed 
to the prefixed) but unfortunately for it to work the 
WebKitAnimationEvent.idl must be backed by a class of the same name 
(at the time of the generation from EventAliases.in the generator is 
not aware yet of implementedAs defined in the IDL of
WebKitAnimationEvent.idl). I also can't alias the prefixed to the 
unprefixed event just like the TransitionEvent because we have to 
honor the runtime flag. When the runtime flag will be removed it 
will be easy to do just like the Transitions and remove the 
WebKit prefix everywhere.

BUG=234610

Review URL: https://chromiumcodereview.appspot.com/22842009

git-svn-id: svn://svn.chromium.org/blink/trunk@156698 bbb929c8-8fbe-4397-9dbb-9b2b20218538