Correctly handle control labels inside canvas fallback content.
The code changes are trivial, just moving code from
AccessibilityRenderObject to AccessibilityNodeObject,
with trivial changes to make the code work there.

Elements inside canvas fallback content are not rendered
but are otherwise accessible; eventually we should move as
much logic as possible to AccessibilityNodeObject so that
there are few differences between inside of a canvas and
outside for accessibility.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/22797007

git-svn-id: svn://svn.chromium.org/blink/trunk@156767 bbb929c8-8fbe-4397-9dbb-9b2b20218538