Web Animations CSS: Implement AnimatableColor
Add support for animating colors in the new web animations engine.
We need to friend RenderStyle to support animating the visited and
unvisited colors separately. Note that we need to update both of them
at least in the case of displaying visited colors and returning
up-to-date unvisited colors to getComputedStyle.

BUG=257591

Review URL: https://chromiumcodereview.appspot.com/23135015

git-svn-id: svn://svn.chromium.org/blink/trunk@156914 bbb929c8-8fbe-4397-9dbb-9b2b20218538