Untangle RenderTheme from Page and RenderObject
Previously, any code wishing to access the theme for a given page had to
go through the Page::theme() method. Because there might not be a page in
some cases, this required NULL checks and/or ASSERTS.

However, having a different RenderTheme based on the Page it's being applied
to is something we're not using in Blink anymore. This also means that the
RenderTheme::defaultTheme() method returns exactly the same as RenderTheme::
themeForPage(), regardless which Page object was being passed in.

Remove the Page::theme() method and change RenderTheme to no longer care
about which page is being passed in. Unify RenderTheme::{themeForPage,
defaultTheme} in RenderTheme::theme(), because the implementation differs
based on the platform. The RenderTheme implementation is now also being
passed as a reference, because it will never be NULL.

BUG=
R=ojan

Review URL: https://chromiumcodereview.appspot.com/23506007

git-svn-id: svn://svn.chromium.org/blink/trunk@156927 bbb929c8-8fbe-4397-9dbb-9b2b20218538