Pre-landing 'binary' file changes for crrev.com/23532016
Both files are UTF-16 encoded and can't be sent through the CQ with
the patch. One has an erroneous stray byte that tickles the new code,
so removing it. And since it's a good idea, the other has an explicit
stray byte to verify the change.

BUG=277035
TBR=jshin@chromium.org

Review URL: https://codereview.chromium.org/23604014

git-svn-id: svn://svn.chromium.org/blink/trunk@157003 bbb929c8-8fbe-4397-9dbb-9b2b20218538