Disable scrollbar hit-test on Android.
We noticed today that slow-path gestures such as taps can still be
forwarded to scrollbars on Android. The Android platform convention is
that scrollbars should not be interactable with directly in any way.
Add a new overlay scrollbar theme parameter to disable hit-testing and
set it on Android.

NOTRY=true
BUG=282704

Review URL: https://chromiumcodereview.appspot.com/23840003

git-svn-id: svn://svn.chromium.org/blink/trunk@157179 bbb929c8-8fbe-4397-9dbb-9b2b20218538