renderStyleOfEnclosingTextNode should update style before using a renderer
When accessing a renderer editing should always make sure that the style
has been updated for the node. Editing calls splitTextNode in a number of
places and then accesses the renderer later. This patch is the first of
many to make sure editing always updates the style before attempting to
access the renderer (or style) of a node after calling a mutation method.

BUG=276689

Review URL: https://chromiumcodereview.appspot.com/23653010

git-svn-id: svn://svn.chromium.org/blink/trunk@157182 bbb929c8-8fbe-4397-9dbb-9b2b20218538